Delete docker socket file only if exists
Using devstack to install kuryr-k8s gets to a point when it breaks as
it's trying to delete the docker socket file, which doesn't exist there yet.
This commit fixes it by having it only delete if it exists.

Closes-Bug: 1685799

+++ devstack/plugin.sh
@@ -276,7 +276,9 @@ function stop_docker {
         echo -n "."
         sleep 1
     done
-    sudo rm "$KURYR_DOCKER_ENGINE_SOCKET_FILE"
+    if [ -e "$KURYR_DOCKER_ENGINE_SOCKET_FILE" ]; then
+        sudo rm "$KURYR_DOCKER_ENGINE_SOCKET_FILE"
+    fi
 }
 
 function get_hyperkube_container_cacert_setup_dir {
